+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.2]
 
 ### Fixes
+- Export default instance from `/default`
 
 ## [1.0.1]